Don't touch GPIO DDR/data on Basic and LF daughterboards by default.
I emailed the mailing list to see if this behavior can be remedied in mainline,
but for now I need to modify it locally.  UHD was driving 0x0000 on the GPIOs
on these boards by default, which could blow up an attached device that expects to
be the only driver.